requirements.txt
click==8.1.3
Flask==2.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
numpy==1.23.0
pandas==1.4.3
protobuf==4.21.2
PyMySQL==1.0.2
python-aqi==0.6.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
xlrd==2.0.1
zipp==3.8.0
